+++ src/components/common/toast/NotificationToast.tsx
@@ -4,6 +4,7 @@ import {
   chatDateFormat,
   notificationDateFormat,
 } from '../../../utils/dateFormat'
+import { typeToColor, typeToIcon } from '../../../utils/typeToCss'
 
 interface NotificationToastType {
   id?: string | number
@@ -12,6 +13,7 @@ interface NotificationToastType {
   content: string //내용은 항상있고
   date: string //알림 들어오는거 보니까 날짜도 있어서
   type: 'chat' | 'notice'
+  notiNode?: string
 }
 
 function NotificationToast({
@@ -21,44 +23,51 @@ function NotificationToast({
   content,
   date,
   type = 'chat',
+  notiNode,
 }: NotificationToastType) {
   //들어오는 타입에 맞춰서 날짜 가공
   const formatDate =
     type === 'chat' ? chatDateFormat(date) : notificationDateFormat(date)
 
   const header = (
-    <div className="flex justify-between border-b border-gray-400 pb-1">
-      <h1>{title}</h1>
+    <div className="flex justify-between border-b border-gray-200 pb-1">
+      <p className="pl-1">{title}</p>
       <button onClick={() => toast.dismiss(id)}>
-        <X className="text-gray-600" />
+        <X className="text-gray-400 hover:text-gray-500 active:text-gray-600" />
       </button>
     </div>
   )
 
   return (
-    <div>
+    <div className="shadow-normal flex w-80 flex-col gap-2 rounded-lg border border-gray-200 bg-white p-2">
       {type === 'chat' ? (
-        <div className="flex w-80 flex-col rounded-sm border border-gray-400 bg-gray-100 p-2">
-          {header}
-          <div className="flex items-center justify-between gap-2">
-            <div>{user}</div>
-            <div className="text-xs text-gray-700">{formatDate}</div>
-          </div>
-          <div className="line-clamp-2 text-sm">{content}</div>
-        </div>
-      ) : (
-        <div className="flex w-80 flex-col rounded-sm border border-gray-400 bg-gray-100 p-2">
+        <div>
           {header}
           <div className="flex items-center justify-between gap-3">
-            <div>
-              {/* 알림 아이콘 만들어진거 있으면 넣어도 괜찮아보여서 추가해둠*/}
-              <Star />
-            </div>
-            <div className="flex grow flex-col">
-              <div className="flex justify-end gap-2 text-xs text-gray-700">
-                {formatDate}
+            <div className="my-3 flex h-full w-full items-center justify-center gap-2.5 pt-1">
+              <div className="flex-1 text-sm">
+                {user}: {content}
               </div>
-              <div className="line-clamp-2 text-sm">{content}</div>
+            </div>
+            <div className="flex h-full gap-3 self-end text-xs whitespace-nowrap text-gray-500">
+              {formatDate}
+            </div>
+          </div>
+        </div>
+      ) : (
+        <div>
+          {header}
+          <div className="flex items-center justify-between gap-3">
+            <div className="my-3 flex h-full w-full items-center justify-center gap-2.5 pt-1">
+              <div
+                className={`flex h-8 w-8 items-center justify-center rounded-full ${notiNode ? typeToColor(notiNode) : 'bg-[#DBEAFE] text-[#2563EB]'}`}
+              >
+                {notiNode ? typeToIcon(notiNode) : <Star />}
+              </div>
+              <div className="flex-1 text-sm">{content}</div>
+            </div>
+            <div className="flex h-full gap-3 self-end text-xs whitespace-nowrap text-gray-500">
+              {formatDate}
             </div>
           </div>
         </div>
