+++ src/types/apiInterface/NotiInterface.ts
@@ -1,5 +1,7 @@
 // 헤더 알림
 
+import type { ReactNode } from 'react'
+
 export interface NotificationParams {
   is_read?: boolean
 }
